+++ core/client/app/templates/subscribers.hbs
@@ -8,7 +8,7 @@
 
     <section class="view-container">
         {{gh-subscribers-table
-                subscribers=sortedSubscribers
+                table=table
                 isLoading=isLoading
                 loadNextPage=(action 'loadNextPage')}}
